--- code/mywebsite.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>About</title>
-    <link href="my.css" rel="stylesheet" type="text/css">
-  </head>
-  <body>
-    <a href="the_game_website.html">our game page</a>
-    <h1>Project Chessman</h1>
-    <p>We're currently making an rpg style game</p>
-    <p> we wouldn't have been able to do this without our mentors AJ and Kris</p>
-    <p>Here's a link to Kris's website Aj doesn't have one so i guess i'll link apple</p>
-    <a href="http://pyxld.com/#/">pyxld.com</a> <a href="https://www.apple.com/">apple.com</a>
-    <p>Programming Team</p>
-    <ul>
-      <li>Alejandro Armas</li>
-      <li>Luis pallares</li>
-    </ul>
-    <p>Art Development Team</p>
-    <ul>
-      <li>Isac Delgado</li>
-      <li>Brad Ayala</li>
-      <li>Eduardo Armas</li>
-      <li>Daniel Aredondo</li>
-      <li>Dekel</li>
-    </ul>
-    <p>Story Development Team</p>
-    <ul>
-      <li>Brad Ayala</li>
-      <li>Isac Delgado</li>
-      <li>Eduardo Armas</li>
-      <li>Daniel Aedondo</li>
-      <li>Dekel<li>
-    </ul>
-  </body>
-</html>
