+++ code/sword.py
@@ -1,0 +1,9 @@
+class Sword(Weapon):
+    def __init__(self, name, damage, image_filename, x, y):
+        Weapon.__init__(self, name, damage, image_filename, x, y)
+        #print("CREATE", self.name)
+    def update():
+        pass
+
+    def draw(self, surface):
+        surface.blit(, )
