--- code/the_game_website.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <link href="my.css" rel="stylesheet" type="text/css">
-</head>
-<body>
-  <h1>Project Chessman</h1>
-  <a href="mywebsite.html">Our About Page</a>
-  <p>Our game us about a young boy who encounters a way to go back in time and meets the aztecs and tries to help them defeat the Spanards</p>
-  <p>Our main charcater</p>
-  <img src="New Piskel clone.png" alt="main charcater" height="400" width="400">
-</body>
-</html>
