+++ code/world.py
@@ -1,0 +1,46 @@
+class World:
+    def __init__(self):
+        self.enemy_list = []
+        self.width = 160
+        self.height = 100
+        self.player = Player(self, 225, 85, 15, 15)
+        self.background = Background(self, pygame.image.load('assets/background_image.png'))
+        self.door = Door(self,self.background, 0, 150)
+        self.door_2 = Door(self,self.background,0,50)
+
+        self.npc = None
+
+
+    def update(self):
+        self.player.update()
+        self.player.Interacting(self.npc)
+        if self.door.check_door_touching(self.player):
+            for enemy_number in range(1, 3):
+                Enemy(self, 50, 50,  pygame.image.load('assets/bad_face_12x12.png'))
+
+
+        #self.door.check_door_touching(self.player)
+        self.player.Interacting(self.npc)
+        for this_enemy in self.enemy_list:
+            this_enemy.update()
+        if self.npc != None:
+            self.npc.update()
+
+    def draw(self):
+        surface.fill((0, 0, 0)) # clear screen (Paint it all white)
+        self.background.draw(surface)
+        self.player.draw(surface)
+        if self.npc != None:
+            self.npc.draw(surface)
+        for this_enemy in self.enemy_list:
+            this_enemy.draw(surface)
+            this_enemy.draw(surface)
+        self.door.draw(surface)
+
+        screen.blit(surface, (0,0)) # this might be what was missing
+        pygame.display.update() #  Actually do all the stuff? (Not actually sure what this does... but I think it should be called at the end of the drawing step)
+        fpsClock.tick(FPS)
+
+    def change_location(self, background_image):
+        self.background.change_image(background_image)
+        self.npc = NonPlayer(self, pygame.image.load('assets/npc_32x32.png'))
