--- js
@@ -1,8 +1,0 @@
-var userInput = 45
-
-if (userInput >= 21){
-  console.console.log();("go get yourself a beer");
-}
-else{
-  console.log("get out of here");
-}
