--- mywebsite
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-  </head>
-  <body>
-    <h1>Project Chessman</h1>
-    <p>We're currently making an rpg style game</p>
-  </body>
-</html>
